+++ test/tools/libtesteth/TestOutputHelper.h
@@ -31,7 +31,7 @@ class TestOutputHelper
 {
 public:
 	TestOutputHelper() { TestOutputHelper::initTest(); }
-	static void initTest(int _maxTests = 1);
+	static void initTest(size_t _maxTests = 1);
 	static void initTest(json_spirit::mValue const& _v);
 	static bool passTest(std::string const& _testName);
 	static void setMaxTests(int _count) { m_maxTests = _count; }
